requirements.txt
Django==2.2.10
djangorestframework==3.12.2
django-nested-admin==3.3.3
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
drf-yasg==1.17.1